+++ tests/testthat/test-kenwardroger.R
@@ -1,131 +1,111 @@
 # h_get_kr_comp ----
 test_that("h_get_kr_comp works as expected on ar1 ungrouped mmrm", {
-    fit <- mmrm(FEV1 ~ ARMCD + ar1(AVISIT | USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
-    expect_snapshot_tolerance(fit$kr_comp)
-  }
-)
+  fit <- mmrm(FEV1 ~ ARMCD + ar1(AVISIT | USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
+  expect_snapshot_tolerance(fit$kr_comp)
+})
 
 test_that("h_get_kr_comp works as expected on ar1 grouped mmrm", {
-    fit <- mmrm(FEV1 ~ ARMCD + ar1(AVISIT | SEX / USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
-    expect_snapshot_tolerance(fit$kr_comp)
-  }
-)
+  fit <- mmrm(FEV1 ~ ARMCD + ar1(AVISIT | SEX / USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
+  expect_snapshot_tolerance(fit$kr_comp)
+})
 
 test_that("h_get_kr_comp works as expected on ar1h ungrouped mmrm", {
-    fit <- mmrm(FEV1 ~ ARMCD + ar1h(AVISIT | USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
-    expect_snapshot_tolerance(fit$kr_comp)
-  }
-)
+  fit <- mmrm(FEV1 ~ ARMCD + ar1h(AVISIT | USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
+  expect_snapshot_tolerance(fit$kr_comp)
+})
 
 test_that("h_get_kr_comp works as expected on ar1h grouped mmrm", {
-    fit <- mmrm(FEV1 ~ ARMCD + ar1h(AVISIT | SEX / USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
-    expect_snapshot_tolerance(fit$kr_comp)
-  }
-)
+  fit <- mmrm(FEV1 ~ ARMCD + ar1h(AVISIT | SEX / USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
+  expect_snapshot_tolerance(fit$kr_comp)
+})
 
 test_that("h_get_kr_comp works as expected on cs ungrouped mmrm", {
-    fit <- mmrm(FEV1 ~ ARMCD + cs(AVISIT | USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
-    expect_snapshot_tolerance(fit$kr_comp)
-  }
-)
+  fit <- mmrm(FEV1 ~ ARMCD + cs(AVISIT | USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
+  expect_snapshot_tolerance(fit$kr_comp)
+})
 
 test_that("h_get_kr_comp works as expected on cs grouped mmrm", {
-    fit <- mmrm(FEV1 ~ ARMCD + cs(AVISIT | SEX / USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
-    expect_snapshot_tolerance(fit$kr_comp)
-  }
-)
+  fit <- mmrm(FEV1 ~ ARMCD + cs(AVISIT | SEX / USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
+  expect_snapshot_tolerance(fit$kr_comp)
+})
 
 test_that("h_get_kr_comp works as expected on csh ungrouped mmrm", {
-    fit <- mmrm(FEV1 ~ ARMCD + csh(AVISIT | USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
-    expect_snapshot_tolerance(fit$kr_comp)
-  }
-)
+  fit <- mmrm(FEV1 ~ ARMCD + csh(AVISIT | USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
+  expect_snapshot_tolerance(fit$kr_comp)
+})
 
 test_that("h_get_kr_comp works as expected on csh grouped mmrm", {
-    fit <- mmrm(FEV1 ~ ARMCD + csh(AVISIT | SEX / USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
-    expect_snapshot_tolerance(fit$kr_comp)
-  }
-)
+  fit <- mmrm(FEV1 ~ ARMCD + csh(AVISIT | SEX / USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
+  expect_snapshot_tolerance(fit$kr_comp)
+})
 
 test_that("h_get_kr_comp works as expected on toep ungrouped mmrm", {
-    fit <- mmrm(FEV1 ~ ARMCD + toep(AVISIT | USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
-    expect_snapshot_tolerance(fit$kr_comp)
-  }
-)
+  fit <- mmrm(FEV1 ~ ARMCD + toep(AVISIT | USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
+  expect_snapshot_tolerance(fit$kr_comp)
+})
 
 test_that("h_get_kr_comp works as expected on toep grouped mmrm", {
-    fit <- mmrm(FEV1 ~ ARMCD + toep(AVISIT | SEX / USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
-    expect_snapshot_tolerance(fit$kr_comp)
-  }
-)
+  fit <- mmrm(FEV1 ~ ARMCD + toep(AVISIT | SEX / USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
+  expect_snapshot_tolerance(fit$kr_comp)
+})
 
 test_that("h_get_kr_comp works as expected on toeph ungrouped mmrm", {
-    fit <- mmrm(FEV1 ~ ARMCD + toeph(AVISIT | USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
-    expect_snapshot_tolerance(fit$kr_comp)
-  }
-)
+  fit <- mmrm(FEV1 ~ ARMCD + toeph(AVISIT | USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
+  expect_snapshot_tolerance(fit$kr_comp)
+})
 
 test_that("h_get_kr_comp works as expected on toeph grouped mmrm", {
-    fit <- mmrm(FEV1 ~ ARMCD + toeph(AVISIT | SEX / USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
-    expect_snapshot_tolerance(fit$kr_comp)
-  }
-)
+  fit <- mmrm(FEV1 ~ ARMCD + toeph(AVISIT | SEX / USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
+  expect_snapshot_tolerance(fit$kr_comp)
+})
 
 test_that("h_get_kr_comp works as expected on us ungrouped mmrm", {
-    fit <- mmrm(FEV1 ~ ARMCD + us(AVISIT | USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
-    expect_snapshot_tolerance(fit$kr_comp)
-  }
-)
+  fit <- mmrm(FEV1 ~ ARMCD + us(AVISIT | USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
+  expect_snapshot_tolerance(fit$kr_comp)
+})
 
 test_that("h_get_kr_comp works as expected on us grouped mmrm", {
-    fit <- mmrm(FEV1 ~ ARMCD + us(AVISIT | SEX / USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
-    expect_snapshot_tolerance(fit$kr_comp)
-  }
-)
+  fit <- mmrm(FEV1 ~ ARMCD + us(AVISIT | SEX / USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
+  expect_snapshot_tolerance(fit$kr_comp)
+})
 
 test_that("h_get_kr_comp works as expected on adh ungrouped mmrm", {
-    fit <- mmrm(FEV1 ~ ARMCD + adh(AVISIT | USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
-    expect_snapshot_tolerance(fit$kr_comp)
-  }
-)
+  fit <- mmrm(FEV1 ~ ARMCD + adh(AVISIT | USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
+  expect_snapshot_tolerance(fit$kr_comp)
+})
 
 test_that("h_get_kr_comp works as expected on adh grouped mmrm", {
-    fit <- mmrm(FEV1 ~ ARMCD + adh(AVISIT | SEX / USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
-    expect_snapshot_tolerance(fit$kr_comp)
-  }
-)
+  fit <- mmrm(FEV1 ~ ARMCD + adh(AVISIT | SEX / USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
+  expect_snapshot_tolerance(fit$kr_comp)
+})
 
 test_that("h_get_kr_comp works as expected on ad ungrouped mmrm", {
-    fit <- mmrm(FEV1 ~ ARMCD + ad(AVISIT | USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
-    expect_snapshot_tolerance(fit$kr_comp)
-  }
-)
+  fit <- mmrm(FEV1 ~ ARMCD + ad(AVISIT | USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
+  expect_snapshot_tolerance(fit$kr_comp)
+})
 
 test_that("h_get_kr_comp works as expected on ad grouped mmrm", {
-    fit <- mmrm(FEV1 ~ ARMCD + ad(AVISIT | SEX / USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
-    expect_snapshot_tolerance(fit$kr_comp)
-  }
-)
+  fit <- mmrm(FEV1 ~ ARMCD + ad(AVISIT | SEX / USUBJID), data = fev_data, reml = TRUE, method = "Kenward-Roger")
+  expect_snapshot_tolerance(fit$kr_comp)
+})
 
 test_that("h_get_kr_comp works as expected on spatial mmrm", {
-    fit <- mmrm(
-      FEV1 ~ ARMCD + sp_exp(VISITN, VISITN2 | USUBJID),
-      data = fev_data, reml = TRUE,
-      method = "Kenward-Roger"
-    )
-    expect_snapshot_tolerance(fit$kr_comp)
-  }
-)
+  fit <- mmrm(
+    FEV1 ~ ARMCD + sp_exp(VISITN, VISITN2 | USUBJID),
+    data = fev_data, reml = TRUE,
+    method = "Kenward-Roger"
+  )
+  expect_snapshot_tolerance(fit$kr_comp)
+})
 
 test_that("h_get_kr_comp works as expected on grouped spatial mmrm", {
-    fit <- mmrm(
-      FEV1 ~ ARMCD + sp_exp(VISITN, VISITN2 | SEX / USUBJID),
-      data = fev_data, reml = TRUE,
-      method = "Kenward-Roger"
-    )
-    expect_snapshot_tolerance(fit$kr_comp)
-  }
-)
+  fit <- mmrm(
+    FEV1 ~ ARMCD + sp_exp(VISITN, VISITN2 | SEX / USUBJID),
+    data = fev_data, reml = TRUE,
+    method = "Kenward-Roger"
+  )
+  expect_snapshot_tolerance(fit$kr_comp)
+})
 
 # df_1d ----
 
@@ -144,7 +124,10 @@ test_that("kr give similar results as SAS for ar1", {
 ## kr linear ----
 
 test_that("kr linear give similar results as SAS for ar1", {
-  fit <- mmrm(FEV1 ~ ARMCD + ar1(AVISIT | USUBJID), data = fev_data, method = "Kenward-Roger-Linear")
+  fit <- mmrm(
+    FEV1 ~ ARMCD + ar1(AVISIT | USUBJID),
+    data = fev_data, method = "Kenward-Roger", vcov = "Kenward-Roger-Linear"
+  )
   res <- df_1d(fit, contrast = c(0, 1))
   expected <- c(0.96058142305176, 188.46934887972)
   expect_equal(res$df, expected[2], tolerance = 1e-4)
@@ -166,7 +149,10 @@ test_that("kr give similar results as SAS for ar1h", {
 ### kr linear ----
 
 test_that("kr linear give similar results as SAS for ar1h", {
-  fit <- mmrm(FEV1 ~ ARMCD + ar1h(AVISIT | USUBJID), data = fev_data, method = "Kenward-Roger-Linear")
+  fit <- mmrm(
+    FEV1 ~ ARMCD + ar1h(AVISIT | USUBJID),
+    data = fev_data, method = "Kenward-Roger", vcov = "Kenward-Roger-Linear"
+  )
   res <- df_1d(fit, contrast = c(0, 1))
   expected <- c(0.75924807546934, 188.225339095373)
   expect_equal(res$df, expected[2], tolerance = 1e-4)
@@ -188,7 +174,10 @@ test_that("kr give similar results as SAS for cs", {
 ### kr linear ----
 
 test_that("kr linear give similar results as SAS for cs", {
-  fit <- mmrm(FEV1 ~ ARMCD + cs(AVISIT | USUBJID), data = fev_data, method = "Kenward-Roger-Linear")
+  fit <- mmrm(
+    FEV1 ~ ARMCD + cs(AVISIT | USUBJID),
+    data = fev_data, method = "Kenward-Roger", vcov = "Kenward-Roger-Linear"
+  )
   res <- df_1d(fit, contrast = c(0, 1))
   expected <- c(0.7964696053595, 177.038485931223)
   expect_equal(res$df, expected[2], tolerance = 1e-4)
@@ -210,7 +199,10 @@ test_that("kr give similar results as SAS for csh", {
 ### kr linear ----
 
 test_that("kr linear give similar results as SAS for csh", {
-  fit <- mmrm(FEV1 ~ ARMCD + csh(AVISIT | USUBJID), data = fev_data, method = "Kenward-Roger-Linear")
+  fit <- mmrm(
+    FEV1 ~ ARMCD + csh(AVISIT | USUBJID),
+    data = fev_data, method = "Kenward-Roger", vcov = "Kenward-Roger-Linear"
+  )
   res <- df_1d(fit, contrast = c(0, 1))
   expected <- c(0.67403858183242, 190.737701349941)
   expect_equal(res$df, expected[2], tolerance = 1e-3)
@@ -232,7 +224,10 @@ test_that("kr give similar results as SAS for adh", {
 ### kr linear ----
 
 test_that("kr linear give similar results as SAS for adh", {
-  fit <- mmrm(FEV1 ~ ARMCD + adh(AVISIT | USUBJID), data = fev_data, method = "Kenward-Roger-Linear")
+  fit <- mmrm(
+    FEV1 ~ ARMCD + adh(AVISIT | USUBJID),
+    data = fev_data, method = "Kenward-Roger", vcov = "Kenward-Roger-Linear"
+  )
   res <- df_1d(fit, contrast = c(0, 1))
   expected <- c(0.66158550758897, 162.393385281755)
   expect_equal(res$df, expected[2], tolerance = 1e-3)
@@ -254,7 +249,10 @@ test_that("kr give similar results as SAS for toep", {
 ### kr linear
 
 test_that("kr linear give similar results as SAS for toep", {
-  fit <- mmrm(FEV1 ~ ARMCD + toep(AVISIT | USUBJID), data = fev_data, method = "Kenward-Roger-Linear")
+  fit <- mmrm(
+    FEV1 ~ ARMCD + toep(AVISIT | USUBJID),
+    data = fev_data, method = "Kenward-Roger", vcov = "Kenward-Roger-Linear"
+  )
   res <- df_1d(fit, contrast = c(0, 1))
   expected <- c(0.87839805519623, 160.027408337368)
   expect_equal(res$df, expected[2], tolerance = 1e-3)
@@ -276,7 +274,10 @@ test_that("kr give similar results as SAS for toeph", {
 ### kr linear
 
 test_that("kr linear give similar results as SAS for toeph", {
-  fit <- mmrm(FEV1 ~ ARMCD + toeph(AVISIT | USUBJID), data = fev_data, method = "Kenward-Roger-Linear")
+  fit <- mmrm(
+    FEV1 ~ ARMCD + toeph(AVISIT | USUBJID),
+    data = fev_data, method = "Kenward-Roger", vcov = "Kenward-Roger-Linear"
+  )
   res <- df_1d(fit, contrast = c(0, 1))
   expected <- c(0.72537324518435, 180.062730071701)
   expect_equal(res$df, expected[2], tolerance = 1e-3)
@@ -302,7 +303,10 @@ test_that("kr give similar results as SAS for unstructured", {
 ### kr linear
 
 test_that("kr linear give similar results as SAS for unstructured", {
-  fit <- mmrm(FEV1 ~ ARMCD + us(AVISIT | USUBJID), data = fev_data, method = "Kenward-Roger-Linear")
+  fit <- mmrm(
+    FEV1 ~ ARMCD + us(AVISIT | USUBJID),
+    data = fev_data, method = "Kenward-Roger", vcov = "Kenward-Roger-Linear"
+  )
   res <- df_1d(fit, contrast = c(0, 1))
   expected <- c(0.66124382270307, 160.73326640376)
   expect_equal(res$df, expected[2], tolerance = 1e-3)
@@ -314,7 +318,10 @@ test_that("kr linear give similar results as SAS for unstructured", {
 ### kr
 
 test_that("kr give similar results as SAS for spatial exponential", {
-  fit <- mmrm(FEV1 ~ ARMCD + sp_exp(VISITN, VISITN2 | USUBJID), data = fev_data, method = "Kenward-Roger-Linear")
+  fit <- mmrm(
+    FEV1 ~ ARMCD + sp_exp(VISITN, VISITN2 | USUBJID),
+    data = fev_data, method = "Kenward-Roger", vcov = "Kenward-Roger-Linear"
+  )
   res <- df_1d(fit, contrast = c(0, 1))
   expected <- c(0.90552903839818, 195.584197921463)
   expect_equal(res$df, expected[2], tolerance = 1e-3)
@@ -324,7 +331,10 @@ test_that("kr give similar results as SAS for spatial exponential", {
 ### kr linear
 
 test_that("kr linear give similar results as SAS for spatial exponential", {
-  fit <- mmrm(FEV1 ~ ARMCD + sp_exp(VISITN, VISITN2 | USUBJID), data = fev_data, method = "Kenward-Roger-Linear")
+  fit <- mmrm(
+    FEV1 ~ ARMCD + sp_exp(VISITN, VISITN2 | USUBJID),
+    data = fev_data, method = "Kenward-Roger", vcov = "Kenward-Roger-Linear"
+  )
   res <- df_1d(fit, contrast = c(0, 1))
   expected <- c(0.90527620094771, 195.584197921463)
   expect_equal(res$df, expected[2], tolerance = 1e-3)
@@ -335,16 +345,16 @@ test_that("kr linear give similar results as SAS for spatial exponential", {
 
 test_that("h_df_1d_kr works as expected in the standard case", {
   object_mmrm_kr <- get_mmrm_kr()
-  expect_snapshot_tolerance(h_df_1d_kr(object_mmrm_kr, c(0, 1), TRUE))
-  expect_snapshot_tolerance(h_df_1d_kr(object_mmrm_kr, c(1, 1), FALSE))
+  expect_snapshot_tolerance(h_df_1d_kr(object_mmrm_kr, c(0, 1)))
+  expect_snapshot_tolerance(h_df_1d_kr(object_mmrm_kr, c(1, 1)))
 })
 
 # h_df_md_kr ----
 
 test_that("h_df_md_kr works as expected in the standard case", {
   object_mmrm_kr <- get_mmrm_kr()
-  expect_snapshot_tolerance(h_df_md_kr(object_mmrm_kr, matrix(c(0, 1, 1, 0), nrow = 2), TRUE))
-  expect_snapshot_tolerance(h_df_md_kr(object_mmrm_kr, matrix(c(0, -1, 1, 0), nrow = 2), FALSE))
+  expect_snapshot_tolerance(h_df_md_kr(object_mmrm_kr, matrix(c(0, 1, 1, 0), nrow = 2)))
+  expect_snapshot_tolerance(h_df_md_kr(object_mmrm_kr, matrix(c(0, -1, 1, 0), nrow = 2)))
 })
 
 # h_kr_df ----