+++ NEWS.md
@@ -6,7 +6,11 @@
   to specifying the model's covariance structure and to expose covariance
   customization through the `tidymodels` interface.
 - Add Kenward-Roger support for spatial covariance structures.
-- Add optional `tidymodels` support.
+- Add empirical and empirical Jackknife adjusted coefficients covariance matrix.
+  In addition, the argument `method` now only specifies the method used
+  for the degrees of freedom, another argument `vcov` is added to specify the
+  method used to adjust the coefficients covariance matrix.
+- Add optional `tidymodels` framework support.
 
 # mmrm 0.2.2